refactor(web): split comment-renderers into per-content-type directory
Each content_type (text/options/preview/trace/system/run/action/credential_request)
now lives in its own file under components/comment-renderers/. The dispatcher
in index.tsx routes through a typed registry indexed on CommentContentType —
adding a new variant becomes a compile error until a renderer is wired.

Public API (CommentRenderer, CommentData, CommentReactions, inlineEventIcon,
isInlineEventType) is unchanged; the existing import path resolves to the new
directory's index.
